[REEF-382] Remove ConfigurationFile
This addressed the issue by
  * replacing serialization usages of ConfigurationFile in our code with AvroConfigurationSerializer
  * moving toConfigurationStringList method to ConfigurationModule
  * deleting ConfigurationFile and TestConfigFileParser

JIRA:
  [REEF-382](https://issues.apache.org/jira/browse/REEF-382)
